*** empty log message ***

+++ Applications/CNRP/include/cnrp_types.h
@@ -88,8 +88,8 @@ typedef struct CNRP_PROBLEM{
 
    best_tours     *cur_tour;  /* temporary tour storage */
    int            *cur_sol_tree;
-   int             fixed_cost;
-   int             variable_cost;
+   double          fixed_cost;
+   double          variable_cost;
    small_graph    *g;         /* contains the edge data for the reduced graph*/
 #if defined(CHECK_CUT_VALIDITY) || defined(TRACE_PATH)
    int             feas_sol_size;
